chore: rename package structure and artifact id
