+++ Provider/src/main/java/com/spotify/confidence/apply/FlagApplier.kt
@@ -1,0 +1,5 @@
+package com.spotify.confidence.apply
+
+interface FlagApplier {
+    fun apply(flagName: String, resolveToken: String)
+}
