+++ Provider/src/main/java/com/spotify/confidence/cache/DiskStorage.kt
@@ -1,0 +1,15 @@
+package com.spotify.confidence.cache
+
+import com.spotify.confidence.client.ResolvedFlag
+import dev.openfeature.sdk.EvaluationContext
+
+interface DiskStorage {
+    fun store(
+        resolvedFlags: List<ResolvedFlag>,
+        resolveToken: String,
+        evaluationContext: EvaluationContext
+    ): CacheData
+    fun read(): CacheData?
+
+    fun clear()
+}
