--- Provider/src/main/java/dev/openfeature/contrib/providers/apply/FlagApplier.kt
@@ -1,5 +1,0 @@
-package dev.openfeature.contrib.providers.apply
-
-interface FlagApplier {
-    fun apply(flagName: String, resolveToken: String)
-}
