--- Provider/src/main/java/dev/openfeature/contrib/providers/client/Clock.kt
@@ -1,20 +1,0 @@
-package dev.openfeature.contrib.providers.client
-
-import java.util.Calendar
-import java.util.Date
-import java.util.TimeZone
-
-interface Clock {
-    fun currentTime(): Date
-
-    class CalendarBacked private constructor() : Clock {
-
-        companion object {
-            fun systemUTC(): Clock = CalendarBacked()
-        }
-
-        override fun currentTime(): Date = Calendar.getInstance().apply {
-            timeZone = TimeZone.getTimeZone("UTC")
-        }.time
-    }
-}
